Better messaging.

+++ scripts/bins_manage.pl
@@ -219,8 +219,10 @@ sub StartJob {
 
 sub ClearAssembly {
     my ($subDir) = @_;
-    print "Clearing assembly work directory for $subDir.\n";
     my $asmDir = "$subDir/Assembly";
-    File::Copy::Recursive::pathempty($asmDir) || die "Could not empty $asmDir: $!";
-    rmdir $asmDir;
+    if (-d $asmDir) {
+        print "Clearing assembly work directory for $subDir.\n";
+        File::Copy::Recursive::pathempty($asmDir) || die "Could not empty $asmDir: $!";
+        rmdir $asmDir;
+    }
 }
